chore: move rule and endpoint reducers
