+++ ui/src/types/stores.ts
@@ -23,8 +23,8 @@ import {OverlayState} from 'src/overlays/reducers/overlays'
 import {AutoRefreshState} from 'src/shared/reducers/autoRefresh'
 import {LimitsState} from 'src/cloud/reducers/limits'
 import {ChecksState} from 'src/checks/reducers'
-import {NotificationRulesState} from 'src/alerting/reducers/notifications/rules'
-import {NotificationEndpointsState} from 'src/alerting/reducers/notifications/endpoints'
+import {NotificationRulesState} from 'src/notifications/rules/reducers'
+import {NotificationEndpointsState} from 'src/notifications/endpoints/reducers'
 import {AlertBuilderState} from 'src/alerting/reducers/alertBuilder'
 
 import {ResourceState} from 'src/types'
